Styling, tidying up components and actions
