+++ muikku-core-plugins/src/main/resources/META-INF/resources/scripts/src/sass/base/colors/main-function.scss
@@ -26,12 +26,14 @@ $color-main-function-button-pill-next-page-background: $color-tertiary;
 $color-main-function-button-pill-next-page-text: $color-tertiary-contrast;
 $color-main-function-button-pill-toolbar-edit-label-background: transparent;
 $color-main-function-button-pill-toolbar-edit-label-text: rgba(0, 0, 0, 0.5);
-$color-main-function-button-pill-toolbar-edit-label-text-hover:  rgba(0, 0, 0, 1);
+$color-main-function-button-pill-toolbar-edit-label-text-hover: rgba(0,0,0, 1);
 $color-main-function-button-pill-toolbar-edit-label-text-active: rgba(0, 0, 0, 1);
 $color-main-function-button-pill-navigation-edit-label-background: transparent;
 $color-main-function-button-pill-navigation-edit-label-text: rgba(0, 0, 0, 0.5);
 $color-main-function-button-pill-navigation-edit-label-text-hover: rgba(0, 0, 0, 1);
 $color-main-function-button-pill-navigation-edit-label-text-active: rgba(0, 0, 0, 1);
+$color-main-function-application-list-item-disabled-text: #b9b7b7;
+$color-main-function-application-list-item-disabled-background: #fbfbfb;
 $color-main-function-application-list-item-open-border: #f0f0f0;
 $color-main-function-application-panel-actions-border: $color-default-separator-border;
 $color-main-function-application-list-item-checkbox-border: #c7c7c7;
