+++ muikku-core-plugins/src/main/resources/META-INF/resources/scripts/src/sass/elements/form-elements.scss
@@ -33,6 +33,7 @@
     height: 1rem;
     justify-content: center;
     margin-right: 5px;
+    padding: 0;
     width: 1rem;
   }
 
@@ -135,7 +136,6 @@
 
 .form-element__copy-workspace-start-date,
 .form-element__copy-workspace-end-date {
-
   .react-datepicker-wrapper {
     display: block;
   }
@@ -185,8 +185,17 @@
   }
 }
 
+.form-element__label {
+  padding: 0 0 5px 0;
+}
+
 .form-element__input {
-  @include prefix(transition, webkit ms o moz, box-shadow 0.2s ease-in-out, border 0.2s ease-in-out);
+  @include prefix(
+    transition,
+    webkit ms o moz,
+    box-shadow 0.2s ease-in-out,
+    border 0.2s ease-in-out
+  );
   -webkit-appearance: none;
   border: solid 2px $color-form-element-border;
   border-radius: 2px;
@@ -310,7 +319,8 @@
   -moz-appearance: none;
   -webkit-appearance: none;
   appearance: none;
-  background: url("data:image/svg+xml;utf8,<svg xmlns='http://www.w3.org/2000/svg' width='15' height='10' fill='skyblue'><polygon points='0,0 3,0 8,7 12,0 15,0 8,10 0,0' /></svg>") no-repeat right 5px center;
+  background: url("data:image/svg+xml;utf8,<svg xmlns='http://www.w3.org/2000/svg' width='15' height='10' fill='skyblue'><polygon points='0,0 3,0 8,7 12,0 15,0 8,10 0,0' /></svg>")
+    no-repeat right 5px center;
   background-color: transparent;
   border: 2px solid $color-form-element-border;
   border-radius: 2px;
@@ -335,7 +345,6 @@
   }
 
   // scss-lint:enable SelectorFormat
-
 }
 
 .form-element__select--matriculation-exam {
